+++ nomad/structs/services.go
@@ -515,6 +515,9 @@ type ConsulConnect struct {
 
 	// SidecarService is non-nil if a service requires a sidecar.
 	SidecarService *ConsulSidecarService
+
+	// SidecarTask is non-nil if sidecar overrides are set
+	SidecarTask *Task
 }
 
 // Copy the stanza recursively. Returns nil if nil.
@@ -526,6 +529,7 @@ func (c *ConsulConnect) Copy() *ConsulConnect {
 	return &ConsulConnect{
 		Native:         c.Native,
 		SidecarService: c.SidecarService.Copy(),
+		SidecarTask:    c.SidecarTask.Copy(),
 	}
 }
 
@@ -595,7 +599,7 @@ func (s *ConsulSidecarService) Equals(o *ConsulSidecarService) bool {
 type ConsulProxy struct {
 	// Upstreams configures the upstream services this service intends to
 	// connect to.
-	Upstreams []*ConsulUpstream
+	Upstreams []ConsulUpstream
 
 	// Config is a proxy configuration. It is opaque to Nomad and passed
 	// directly to Consul.
@@ -611,10 +615,10 @@ func (p *ConsulProxy) Copy() *ConsulProxy {
 	newP := ConsulProxy{}
 
 	if n := len(p.Upstreams); n > 0 {
-		newP.Upstreams = make([]*ConsulUpstream, n)
+		newP.Upstreams = make([]ConsulUpstream, n)
 
 		for i := range p.Upstreams {
-			newP.Upstreams[i] = p.Upstreams[i].Copy()
+			newP.Upstreams[i] = *p.Upstreams[i].Copy()
 		}
 	}
 
@@ -643,7 +647,7 @@ func (p *ConsulProxy) Equals(o *ConsulProxy) bool {
 OUTER:
 	for _, up := range p.Upstreams {
 		for _, innerUp := range o.Upstreams {
-			if up.Equals(innerUp) {
+			if up.Equals(&innerUp) {
 				// Match; find next upstream
 				continue OUTER
 			}
